+++ src/woocommerce-payment-gateway-cloud.php
@@ -36,7 +36,8 @@ add_action('plugins_loaded', function () {
         return $methods;
     }, 0);
 
-    add_filter('woocommerce_checkout_before_customer_details', function(){
+    // add_filter('woocommerce_before_checkout_form', function(){
+    add_filter('the_content', function(){
         if(!empty($_GET['gateway_return_result']) && $_GET['gateway_return_result'] == 'error') {
             wc_print_notice(__('Payment failed or was declined', 'woocommerce'), 'error');
         }